refactor(shared-feature): move shared features to separated kind
