+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/provide-access-token-response-extractors.ts
@@ -1,13 +1,10 @@
-import {
-  EnvironmentProviders,
-  Provider,
-  Type,
-  makeEnvironmentProviders,
-} from '@angular/core';
+import { EnvironmentProviders, makeEnvironmentProviders } from '@angular/core';
 
-import { generateAccessTokenResponseExtractorProvider } from './provide-helper.functions';
+import {
+  AccessTokenResponseExtractorFeature,
+  SharedProviderFeatureKind,
+} from './provide-shared.features';
 import { ACCESS_TOKEN_RESPONSE_EXTRACTOR_INFOS } from './tokens';
-import { AccessTokenResponse, AccessTokenResponseExtractor } from './types';
 
 export function provideAccessTokenResponseExtractors(
   ...features: AccessTokenResponseExtractorsFeatures[]
@@ -16,52 +13,15 @@ export function provideAccessTokenResponseExtractors(
     .filter(
       (feature): feature is AccessTokenResponseExtractorFeature =>
         feature.kind ===
-        AccessTokenResponseExtractorsFeatureKind.AccessTokenResponseExtractorFeature,
+        SharedProviderFeatureKind.AccessTokenResponseExtractorFeature,
     )
-    .forEach(
-      generateAccessTokenResponseExtractorProvider(
-        ACCESS_TOKEN_RESPONSE_EXTRACTOR_INFOS,
-      ),
+    .forEach((feature) =>
+      feature.assign(ACCESS_TOKEN_RESPONSE_EXTRACTOR_INFOS),
     );
 
   return makeEnvironmentProviders([
     features.map((feature) => feature.providers),
   ]);
-}
-
-export enum AccessTokenResponseExtractorsFeatureKind {
-  // NOTE: This is a shared feature. Make it be unquie.
-  AccessTokenResponseExtractorFeature = 'access-token-response-extractor-feature',
-}
-
-export interface AccessTokenResponseExtractorsFeature<
-  K extends AccessTokenResponseExtractorsFeatureKind,
-> {
-  readonly kind: K;
-  readonly providers: Provider[];
-}
-
-export type AccessTokenResponseExtractorFeature<
-  T extends AccessTokenResponse = AccessTokenResponse,
-  C = unknown,
-> = AccessTokenResponseExtractorsFeature<AccessTokenResponseExtractorsFeatureKind.AccessTokenResponseExtractorFeature> & {
-  type: Type<AccessTokenResponseExtractor<T, C>>;
-  config: C;
-};
-
-export function withAccessTokenResponseExtractor<
-  T extends AccessTokenResponse,
-  C,
->(
-  type: Type<AccessTokenResponseExtractor<T, C>>,
-  config: C,
-): AccessTokenResponseExtractorFeature<T, C> {
-  return {
-    kind: AccessTokenResponseExtractorsFeatureKind.AccessTokenResponseExtractorFeature,
-    type: type,
-    config: config,
-    providers: [],
-  };
 }
 
 export type AccessTokenResponseExtractorsFeatures =
